requirements.txt
Django==1.11.9
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
Faker==0.8.8
